Refactoring for better packaging, Java8 support and minor performance fix.
