--- src/main/java/com/flipkart/zjsonpatch/CopyingApplyProcessor.java
@@ -1,10 +1,0 @@
-package com.flipkart.zjsonpatch;
-
-import com.fasterxml.jackson.databind.JsonNode;
-
-class CopyingApplyProcessor extends InPlaceApplyProcessor {
-
-    CopyingApplyProcessor(JsonNode target) {
-        super(target.deepCopy());
-    }
-}
