--- src/main/java/com/flipkart/zjsonpatch/InvalidJsonPatchException.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2016 flipkart.com zjsonpatch.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-package com.flipkart.zjsonpatch;
-
-/**
- * User: holograph
- * Date: 03/08/16
- */
-public class InvalidJsonPatchException extends JsonPatchApplicationException {
-    public InvalidJsonPatchException(String message) {
-        super(message);
-    }
-
-    public InvalidJsonPatchException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public InvalidJsonPatchException(Throwable cause) {
-        super(cause);
-    }
-}
