+++ src/main/java/com/flipkart/zjsonpatch/constants/Constants.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2016 flipkart.com zjsonpatch.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+package com.flipkart.zjsonpatch.constants;
+
+/**
+ * Created with IntelliJ IDEA.
+ * User: gopi.vishwakarma
+ * Date: 10/07/15
+ * Time: 10:35 AM
+ */
+public final class Constants {
+    public static String OP = "op";
+    public static String VALUE = "value";
+    public static String PATH = "path";
+    public static String FROM = "from";
+
+    private Constants() {}
+
+}
