+++ src/main/java/com/flipkart/zjsonpatch/constants/DiffFlags.java
@@ -1,0 +1,11 @@
+package com.flipkart.zjsonpatch.constants;
+
+import java.util.EnumSet;
+
+public enum DiffFlags {
+    OMIT_VALUE_ON_REMOVE;
+
+    public static EnumSet<DiffFlags> defaults() {
+        return EnumSet.of(OMIT_VALUE_ON_REMOVE);
+    }
+}
