+++ src/test/java/com/flipkart/zjsonpatch/CopyingApplyProcessor.java
@@ -1,0 +1,10 @@
+package com.flipkart.zjsonpatch;
+
+import com.fasterxml.jackson.databind.JsonNode;
+
+class CopyingApplyProcessor extends InPlaceApplyProcessor {
+
+    CopyingApplyProcessor(JsonNode target) {
+        super(target.deepCopy());
+    }
+}
